+++ library/QATools/QATools/PageObject/Element/WebElement.php
@@ -70,7 +70,6 @@ use Behat\Mink\Session;
  * @method boolean hasTable($locator) Checks whether element has a table with specified locator.
  * @method void attachFileToField($locator, $path) Attach file to file field with specified locator.
  *
- * @method Session getSession() Returns element session.
  * @method boolean has($selector, $locator) Checks whether element with specified selector exists inside the current element.
  * @method boolean isValid() Checks if an element still exists in the DOM.
  * @method string getText() Returns element text (inside tag).
@@ -88,6 +87,13 @@ class WebElement implements IWebElement, INodeElementAware
 	private $_wrappedElement;
 
 	/**
+	 * Stores instance of used page factory.
+	 *
+	 * @var IPageFactory
+	 */
+	private $_pageFactory;
+
+	/**
 	 * The XPath escaper.
 	 *
 	 * @var Escaper
@@ -97,11 +103,13 @@ class WebElement implements IWebElement, INodeElementAware
 	/**
 	 * Initializes web element.
 	 *
-	 * @param NodeElement $wrapped_element Wrapped element.
+	 * @param NodeElement  $wrapped_element Wrapped element.
+	 * @param IPageFactory $page_factory    Page factory.
 	 */
-	public function __construct(NodeElement $wrapped_element)
+	public function __construct(NodeElement $wrapped_element, IPageFactory $page_factory)
 	{
 		$this->_wrappedElement = $wrapped_element;
+		$this->_pageFactory = $page_factory;
 		$this->_xpathEscaper = new Escaper();
 	}
 
@@ -113,9 +121,9 @@ class WebElement implements IWebElement, INodeElementAware
 	 *
 	 * @return static
 	 */
-	public static function fromNodeElement(NodeElement $node_element, IPageFactory $page_factory = null)
+	public static function fromNodeElement(NodeElement $node_element, IPageFactory $page_factory)
 	{
-		return new static($node_element);
+		return new static($node_element, $page_factory);
 	}
 
 	/**
@@ -186,6 +194,32 @@ class WebElement implements IWebElement, INodeElementAware
 	public function getXpathEscaper()
 	{
 		return $this->_xpathEscaper;
+	}
+
+	/**
+	 * Returns element session.
+	 *
+	 * @return     Session
+	 * @deprecated Accessing the session from the element is deprecated as of 1.2 and will be impossible in 2.0.
+	 */
+	public function getSession()
+	{
+		@trigger_error(
+			sprintf('The method %s is deprecated as of 1.2 and will be removed in 2.0', __METHOD__),
+			E_USER_DEPRECATED
+		);
+
+		return $this->_pageFactory->getSession();
+	}
+
+	/**
+	 * Returns page factory, used during object creation.
+	 *
+	 * @return IPageFactory
+	 */
+	protected function getPageFactory()
+	{
+		return $this->_pageFactory;
 	}
 
 	/**